Fix the implementation of expressions in TMDL and SAN
- Fix wrong return type for the 'evaluate()' operation in
  TMDL::Expression package.
- Fix implementation of 'ThisContextIterator' in SANT

+++ plugins/org.modelspartiti.infrastructure/src/tmdl/expressions/TemplateExpressionReal.java
@@ -32,6 +32,6 @@ public interface TemplateExpressionReal extends TemplateExpression {
 	 *        annotation="http://www.eclipse.org/emf/2002/Ecore/OCL/Pivot body='evaluate(core::Assignment.allInstances())'"
 	 * @generated
 	 */
-	int evaluate();
+	double evaluate();
 
 } // TemplateExpressionReal
