+++ plugins/org.modelspartiti.infrastructure/src/tmdl/expressions/impl/TemplateExpressionArrayAggregateImpl.java
@@ -186,9 +186,9 @@ public class TemplateExpressionArrayAggregateImpl extends MinimalEObjectImpl.Con
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public int evaluate() {
+	public double evaluate() {
 		try {
-			return (Integer)EVALUATE__EINVOCATION_DELEGATE.dynamicInvoke(this, null);
+			return (Double)EVALUATE__EINVOCATION_DELEGATE.dynamicInvoke(this, null);
 		}
 		catch (InvocationTargetException ite) {
 			throw new WrappedException(ite);
