+++ plugins/org.modelspartiti.infrastructure/src/tmdl/expressions/impl/TemplateExpressionParameterArrayValueAtImpl.java
@@ -199,9 +199,9 @@ public class TemplateExpressionParameterArrayValueAtImpl extends MinimalEObjectI
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public int evaluate() {
+	public double evaluate() {
 		try {
-			return (Integer)EVALUATE__EINVOCATION_DELEGATE.dynamicInvoke(this, null);
+			return (Double)EVALUATE__EINVOCATION_DELEGATE.dynamicInvoke(this, null);
 		}
 		catch (InvocationTargetException ite) {
 			throw new WrappedException(ite);
